Move the BitTorrent related code from SiteAnnouncer.py to a separate plugin
